# src/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pathofmodifiers",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/api_v1/modifier/{modifierId}": {
      "get": {
        "tags": [
          "modifiers"
        ],
        "summary": "Get Modifier",
        "description": "Get modifier or list of modifiers by key and\nvalue for \"modifierId\"\n\nDominant key is \"modifierId\".\n\nReturns one or a list of modifiers.",
        "operationId": "get_modifier",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "modifierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Modifierid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Modifier"
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Modifier"
                      }
                    }
                  ],
                  "title": "Response Modifiers-Get Modifier"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "modifiers"
        ],
        "summary": "Delete Modifier",
        "description": "Delete a modifier by key and value for \"modifierId\"\n\nReturns a message that the modifier was deleted.\nAlways deletes one modifier.",
        "operationId": "delete_modifier",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "modifierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Modifierid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Modifiers-Delete Modifier"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api_v1/modifier/": {
      "get": {
        "tags": [
          "modifiers"
        ],
        "summary": "Get All Modifiers",
        "description": "Get all modifiers.\n\nReturns a list of all modifiers.",
        "operationId": "get_all_modifiers",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Modifier"
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Modifier"
                      }
                    }
                  ],
                  "title": "Response Modifiers-Get All Modifiers"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "modifiers"
        ],
        "summary": "Create Modifier",
        "description": "Create one or a list of new modifiers.\n\nReturns the created modifier or list of modifiers.",
        "operationId": "create_modifier",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "return_nothing",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return Nothing"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ModifierCreate"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ModifierCreate"
                    }
                  }
                ],
                "title": "Modifier"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ModifierCreate"
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ModifierCreate"
                      }
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Modifiers-Create Modifier"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "modifiers"
        ],
        "summary": "Update Modifier",
        "description": "Update a modifier by key and value for \"modifierId\"\n\nDominant key is \"modifierId\".\n\nReturns the updated modifier.",
        "operationId": "update_modifier",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "modifierId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Modifierid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifierUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Modifier"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api_v1/modifier/grouped_modifiers_by_effect/": {
      "get": {
        "tags": [
          "modifiers"
        ],
        "summary": "Get Grouped Modifier By Effect",
        "description": "Get all grouped modifiers by effect.\n\nReturns a list of all grouped modifiers by effect.",
        "operationId": "get_grouped_modifier_by_effect",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/GroupedModifierByEffect"
                    },
                    {
                      "items": {
                        "$ref": "#/components/schemas/GroupedModifierByEffect"
                      },
                      "type": "array"
                    }
                  ],
                  "title": "Response Modifiers-Get Grouped Modifier By Effect"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/api_v1/test/bulk-insert-test": {
      "post": {
        "tags": [
          "tests"
        ],
        "summary": "Bulk Insert Test",
        "description": "Can only be used in `settings.ENVIRONMENT=local` environment.\n\nTest route for bulk inserting records.\n\nReturns a success message once the insertion is complete.",
        "operationId": "bulk_insert_test",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "count",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Tests-Bulk Insert Test"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api_v1/test/bulk-insert-users-and-verify": {
      "post": {
        "tags": [
          "tests"
        ],
        "summary": "Bulk Insert Users And Verify",
        "description": "Can only be used in `settings.ENVIRONMENT=local` environment.\n\nTest route for bulk inserting users and verifying them.\n\nReturns the access tokens for the created users.",
        "operationId": "bulk_insert_users_and_verify",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "count",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Tests-Bulk Insert Users And Verify"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GroupedModifierByEffect": {
        "properties": {
          "effect": {
            "type": "string",
            "title": "Effect"
          },
          "regex": {
            "type": "string",
            "title": "Regex"
          },
          "static": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Static"
          },
          "relatedUniques": {
            "type": "string",
            "title": "Relateduniques"
          },
          "groupedModifier": {
            "$ref": "#/components/schemas/GroupedModifier"
          }
        },
        "type": "object",
        "required": [
          "effect",
          "regex",
          "static",
          "relatedUniques",
          "groupedModifier"
        ],
        "title": "GroupedModifierByEffect"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Modifier": {
        "properties": {
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "relatedUniques": {
            "type": "string",
            "title": "Relateduniques"
          },
          "minRoll": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minroll"
          },
          "maxRoll": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxroll"
          },
          "textRolls": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Textrolls"
          },
          "static": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Static"
          },
          "effect": {
            "type": "string",
            "title": "Effect"
          },
          "regex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regex"
          },
          "implicit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implicit"
          },
          "explicit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explicit"
          },
          "delve": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delve"
          },
          "fractured": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fractured"
          },
          "synthesized": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthesized"
          },
          "unique": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unique"
          },
          "corrupted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrupted"
          },
          "enchanted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enchanted"
          },
          "veiled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Veiled"
          },
          "modifierId": {
            "type": "integer",
            "title": "Modifierid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "position",
          "relatedUniques",
          "effect",
          "modifierId",
          "createdAt"
        ],
        "title": "Modifier"
      },
      "ModifierCreate": {
        "properties": {
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "relatedUniques": {
            "type": "string",
            "title": "Relateduniques"
          },
          "minRoll": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minroll"
          },
          "maxRoll": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxroll"
          },
          "textRolls": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Textrolls"
          },
          "static": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Static"
          },
          "effect": {
            "type": "string",
            "title": "Effect"
          },
          "regex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regex"
          },
          "implicit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implicit"
          },
          "explicit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explicit"
          },
          "delve": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delve"
          },
          "fractured": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fractured"
          },
          "synthesized": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthesized"
          },
          "unique": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unique"
          },
          "corrupted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrupted"
          },
          "enchanted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enchanted"
          },
          "veiled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Veiled"
          }
        },
        "type": "object",
        "required": [
          "position",
          "relatedUniques",
          "effect"
        ],
        "title": "ModifierCreate"
      },
...
      "ModifierUpdate": {
        "properties": {
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "relatedUniques": {
            "type": "string",
            "title": "Relateduniques"
          },
          "minRoll": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minroll"
          },
          "maxRoll": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxroll"
          },
          "textRolls": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Textrolls"
          },
          "static": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Static"
          },
          "effect": {
            "type": "string",
            "title": "Effect"
          },
          "regex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regex"
          },
          "implicit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implicit"
          },
          "explicit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explicit"
          },
          "delve": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delve"
          },
          "fractured": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fractured"
          },
          "synthesized": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthesized"
          },
          "unique": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unique"
          },
          "corrupted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrupted"
          },
          "enchanted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enchanted"
          },
          "veiled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Veiled"
          }
        },
        "type": "object",
        "required": [
          "position",
          "relatedUniques",
          "effect"
        ],
        "title": "ModifierUpdate"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/api_v1/login/access-token"
          }
        }
      }
    }
  }
}